fix(Pilot Sheet): added a Pilot Roster reorder/lock order toggle to touch devices
vuetify seems to be overriding an event trigger that lets draggable determine if the touch event is
a click or the start of a touch/drag. Short of starting to look for a new dragging library or
rewriting the roster page, I added a drag lock toggle that should only appear on touch-capable
devices.

Closes #896