Always update the NPC roster when NPCs are loaded
There was a race condition in the async loading of this array. Sometimes the Startup() method wouldn't be called on page refresh. Sometimes the LOAD_NPCS call wouldn't complete until after the NpcManager had already been created. This code replaced this.Npcs rather than updating it, so reactivity couldn't take care of the race condition.