Fix inconsistent behavior in NPC importing
Setting the value that was the v-model during @change was causing this method to be called over and over again. It would eventually fail because the file object passed in was invalid. This changes it to use v-model and a watch to get better consistency on when the file import is called. This will also reset the file when we're done uploading.