fix(Pilot Sheet): new Prototype Weapon dynamically updates uses + damage
New MechWeapon objects did not set a default value for `max_use_override` or `_custom_damage_type`
upon creation; this prevented the weapon card from grabbing those fields to update the UI.  The
mentioned fields *are* set to a default when deserialized on reload, which is why refreshing the
page made uses/damage dynamic again.
Additional change included to prevent negative max uses or
enough max uses to overflow the page; current max uses is 25 but can be modified if appropriate in
the future.

Closes #1533