fix(Mech Sheet): frames with webhosted images use the webhost address
With this fix, in the image selector, if the selected image's source is a URL, it uses the image URL
for displaying the image.
Prior to this fix, the selected image would only have its basename stored
as a "local" filename, which appended `static/img/(mech/pilot)` to the file's basename.  This led to
broken references for frames that lacked a corresponding local static image file, e.g. Wallflower
frames and Long Rim frames with alternative naming schemes (e.g. `lich_off.png` remote vs.
`mf_lich.png` local).

Closes  #1439