fix(Pilot Sheet): fix Skill Focus, setup custom reserve retrieval
This PR started with fixing an issue introduced by #1564, and developed into realizing we don't
actually store bonuses/actions/synergies/etc. for custom-made reserves.  Any bonuses, actions,
synergies, etc. would only be available for the Reserves listed in `lancer-data`, as reserves with
IDs matching those in `lancer-data` are retrieved fresh every time.  This PR aims to address the
discrepancy, and hopefully will open the way to a wider variety of custom reserves.

Closes #1589