fix(Mech Sheet): scale stats with max stat increases
Use a "missing stat" approach similar to the HP changes in #1569 for similar statistics so that the
"current" values scale with increases to the maximum values.  Most noticeable with Repair Cap due to
Hull bonuses; Structure & Stress included in case respective bonuses are introduced.  Also performed
similar scaling for Limited-use systems (though more of the math was on the Vue/HTML side), most
noticeable on Prototype weapon.

Closes #1586