fix(Pilot Sheet): enable Pilot Roster reordering within groups
Previously, Pilot Roster was sortable by groups, but pilots couldn't be sorted within those groups;
they were always displayed in the order of creation (and this order was never modified).  This PR
aims to move some of the Pilot Roster management out of the Vue component and into the
PilotManagementStore: Adding, deleting, & renaming groups are incorporated into the store, as well
as sorted loading based upon Pilot Group & moving/reindexing the Pilot array on click & drag.

Closes #1625