fix(Mech Sheet): handle mechs with frame sizes greater than 3
This change modifies the logic behind a Mech's size calculation.  It returns the size of the mech's
Frame, unaltered, if the Frame's size already equals or exceeds the max frame size set by the rules.
Otherwise, it includes any bonuses to the mech's size, as normal.  This primarily benefits any
homebrew content making Size 4 frames.

Closes #1682