style(Compendium): prevent overcropping License frame art
This change increases visibility of the frame art in the License Compendium.  Prior to this,
expanding the viewable window led to a larger right-side cropping of the frame art.  While the
license banner loses some balance as a result, it will lead to fewer "weird" crops of the frames'
art.

Closes #1690