fix(Pilot Sheet): improve Vault Code UX
This change aims to streamline the process of acquiring a pilot's vault code. It does so by limiting
Vault Code access to when the user is logged in, improving messaging when a pilot has yet to sync,
and detecting/updating a pilot's cloud ID information once a sync has occurred.

Closes #1664