feat(Pilot Sheet): allow skipping pilot creation
This PR subtly enables the creation of a new pilot even when missing parameters from earlier in the
wizard.  It assigns a default name and callsign for ease of reading.  This approach leans on the
subtle side to encourage users to follow the wizard, but still allows experienced users to quickly
create a new pilot to edit directly in Pilot management.

Closes #1687