fix(Mech Sheet): remove "ghost" aux slot from Flex slot
This change revisits the previous bugfix for extra Auxillary weapons in the Flex slot.  Previous
version didn't have a safety check to see if the `extra` slot was already empty.  This led to a
circular call of UnequipWeapon -> get Slots -> ad infinitum.  Not immediately noticable on local dev
unless monitoring performance (or dropping a logging statement in the `else if` clause).  This
version should safely unequip the "ghost" Aux slots without going infinite.

Closes #1692