fix(Active Mode): limited uses initialize to correct maximum
This change retools how Max Uses and Total Limited Uses are calculated and presented in Active mode
to ensure that a mech has the correct number of limited uses at the start of a mission.  It changes
the logic behind Max Use Override so that bonuses must be manually included in the override. Lastly,
this includes a couple of other minor changes that polish some of the Limited Systems behaviors in
Active Mode.

Closes #1708