feat(Cloud): settings for manually pulling data from cloud
This change separates the "On App Load" and "On User Sign In" options into a "Load Frequency"
category, indicating that when enabled, CompCon will automatically pull from the cloud when those
events trigger.  This change also adds a "Cloud Load" button that manually pulls data from the
cloud.  Lastly, this change modifies wording surrounding sync functions to clarify what is occurring
during the sync (a "save" to the cloud, or a "load" from the cloud).

Closes #1731