feat(Pilot Sheet): initial pilot group draggable changes
This is an initial commit for improving performance in the Pilot Roster.  It lays the groundwork for
using nested Draggable components, utilizing the yet-unused `pilot_groups.json` for persisting
Roster organization, and fixes autoscrolling issues (and drag-and-drop ghost issues in Chrome).  It
is far from complete; PilotStore needs cleaned up to fit the new Group paradigm, the Pilot Groups
object needs to be used directly in Draggable, and it should probably be load-tested to see if it
can handle large lists of pilots.