fix(Pilot Sheet): fixes vault import bug for pilot group update
This is a fix for the previous Pilot Group update.  It disables the renewal of a new pilot ID upon
calling `SetRemoteResource`, which led to ID mismatches between the Pilot Group structure and the
Pilot itself.  Weirdly enough, the Renewed ID was stored in the Pilot Group structure, while the
Pilot's ID reverted back to the Pilot's CloudID. It's not clear why that reversion occurred, but it
makes sense to me to keep the Pilot's ID the same as the CloudID instead of renewing it in these
cases.  Feedback/explanations for the contrary would be appreciated!

This change also prevents importing pilots from the user's own account, as this leads to strange behavior
when modifying/deleting the pilot.  Better to prevent it in the first place.  Also adds more informative
error messages when a pilot ID is not found.

Closes #1777