fix(accessibility): fix compendium button not showing focus
This is a questionable fix.
v-btns that contain a link (rather than an onclick function, for example) implicitly use an anchor tag instead of a button. This causes differing styling, including missing focus animation.
I'm working around this by removing the `to` attribute from the v-button wrapping the button in an anchor that handles the navigation.
A more elegant solution may be subclassing v-btn to always provide this functionality.

wrap compendium nav button in anchor