Unrestricted Weapon Mods
When unrestricted weapon toggle is active the regular checks are bypassed for the filtering in the _ModSelector. This allows users to select mods that would normally be incompatible that they have access to.

To inform the user that such a choice has been made, an additional status alert has been created to alert the player when modifying the mech in the Mech Hanger and at the Start Mission stage of the Pilot Roster