Fix mission runner page refresh race condition.
The Vue container in Active.vue was trying to use the MissionStore
before it fully loaded, resulting in it being undefined.  This adds a
condition on the container to see if there is an active mission that is
truthy and displays the content if it is.

If there are active missions in the store, but there are none found
(e.g. the UUID for the active mission is invalid in the route path),
then it will display an error message.