Fix dupe integrated weapons when granted by CBs
It turns out that since CBS are applied to both Pilots & Mechs, the code collecting IWs/ISs can and will accidentally duplicate them. In lieu of refactoring the collectors to take a context or something, I've made the executive decision to just... not support adding the same thing as an integrated weapon bonus twice. This should have zero negative effects, since integrated gear is generally supposed to be entirely unique (and therefore there's no other second thing to apply it). I'm tentatively open to revisiting this if there's a genuine need for it that can't be solved with "This weapon can be attacked with twice in a Barrage" or something similar, though.