fix: properly filter exotic mods
previously, exotic mods were not filtered from the mod selector menu when not added to a pilot as special equipment, and were in the list regardless of compatibility. this pr causes them to behave as any other exotic item, only appearing when they are added to a pilot and when compatible (unless show incompatible is toggled).